feat: 디바이스 등록 API 연결 - DTO, 서비스, 뷰모델 전부 수정 - HomeScreen 37번째 줄 추가
Jira: feature/mobile/design-S13P21D208-240

+++ frontend/app/src/main/java/com/example/eeum/data/model/response/floorplans/FloorPlansList.kt
@@ -3,8 +3,8 @@ package com.example.eeum.data.model.response.floorplans
 data class FloorPlansList(
     val createdAt: String,
     val floorplanId: Int,
-    val floorplansX: Int,
-    val floorplansY: Int,
+    val floorplansX: Float,
+    val floorplansY: Float,
     val homeId: Int,
     val imageUrl: String,
     val square: Double,
